+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [1.0.0.48] - 2026-09-20
+
 ### Added
 
 - A per-client circuit breaker: `RipClientConfig.Builder#circuitBreaker(CircuitBreakerConfig)`